backend/requirements.txt
requests
python-dotenv

# Database - PostgreSQL with pgvector
# Note: pgvector is a PostgreSQL extension. Install it in your database:
#   CREATE EXTENSION vector;
# For Python, we use psycopg2-binary to connect to PostgreSQL
psycopg2-binary

# Document Processing
pdfplumber
PyPDF2
Pillow
pytesseract

# Embeddings and AI - Azure AI Foundry Service
openai
azure-identity
azure-core
numpy

# Azure Computer Vision - Image Analysis 4.0 API for READ feature
azure-ai-vision-imageanalysis

# Vector Database - Pinecone (Direct API)

# LangGraph and LangChain (for agent workflows)
langgraph
langchain
langchain-openai
langchain-core
langchain-community

# Unstructured document processing (for PDF loading)
unstructured[pdf]

# Medical coding and claims processing
# EDI processing (for HIPAA 837, 270/271 transactions)
# Note: For production, consider using python-edi or pyx12 for proper EDI handling
# For now, we use basic string formatting

# Additional utilities
python-dateutil
